Refactor to cleanup discovery for Local HW
